Nat

Description

Documentation

data Nat

Datatype representing the Natural numbers as Zero and Succ(essors) of Zero.

0 = Zero 1 = Succ Zero 2 = Succ (Succ Zero)

Constructors

Succ Nat

Zero

intToNat :: Int -> Nat

Builds a Nat from an integer, linear on the size of the Int.

Has a lower bound of Zero.

natAdd :: Nat -> Nat -> Nat

Nat + Nat operation.

Accumulates Succ(essors) from first digit onto the other, so this operation is linear on the size of the first digit unless the second is Zero.

natToInt :: Nat -> Int

Builds an integer from a Nat, linear on the size of the Nat.

natPred :: Nat -> Nat

Nat - 1 operation, with a lower bound of Zero.