Term

Description

Documentation

data Term

Untyped lambda terms using De Bruijn indices to avoid variable capture

Constructors

Var Nat

Lam Term

App Term Term

type ReductionSequence = [Term]

Provides extra context to specific functions dealing with term lists

termListToStr :: ReductionSequence -> String

Pretty prints a reduction sequence of terms